+++ lib/routes/cnbeta/common.ts
@@ -1,4 +1,3 @@
-import { Route } from '@/types';
 import cache from '@/utils/cache';
 import got from '@/utils/got';
 import { load } from 'cheerio';
@@ -7,21 +6,7 @@ import { parseDate } from '@/utils/parse-date';
 
 import { rootUrl, ProcessItems } from './utils';
 
-export const route: Route = {
-    path: ['/:type/:id', '/'],
-    radar: [
-        {
-            source: ['cnbeta.com.tw/'],
-            target: '',
-        },
-    ],
-    name: 'Unknown',
-    maintainers: [],
-    handler,
-    url: 'cnbeta.com.tw/',
-};
-
-async function handler(ctx) {
+export async function handler(ctx) {
     const { type, id } = ctx.req.param();
 
     const currentUrl = type ? `${rootUrl}/${type}/${id}.htm` : rootUrl;
