+++ lib/routes/cnbeta/namespace.ts
@@ -3,4 +3,5 @@ import type { Namespace } from '@/types';
 export const namespace: Namespace = {
     name: 'cnBeta.COM',
     url: 'cnbeta.com.tw',
+    categories: ['new-media'],
 };
