+++ lib/routes/dnaindia/namespace.ts
@@ -3,4 +3,5 @@ import type { Namespace } from '@/types';
 export const namespace: Namespace = {
     name: 'DNA India',
     url: 'dnaindia.com',
+    categories: ['traditional-media'],
 };
